# api/openapi.yaml
openapi: 3.0.3
info:
  title: Calendar API
  version: 1.0.0
  contact:
    name: Source code
    url: https://github.com/Alexander-D-Karpov/calendar
  description: |
    JSON API for calendars, events, todos and share links. All operations live under `/api/v1`.

    ## Authentication

    Send an API token in the `Authorization` header:

        Authorization: Bearer cal_<prefix>_<secret>

    Create tokens in [Settings → API tokens](/settings/tokens). A token is shown once, right after it is created, and can be revoked at any time. To try requests on this page with a token, press **Authorize** and paste it.

    Requests from a signed-in browser session work too. With a session, unsafe methods (`POST`, `PUT`, `PATCH` and `DELETE`) also need the `X-CSRF-Token` header. **Try it out** on this page adds that header when you are signed in.

    ## Scopes

    Each token carries one or more scopes, and each operation names the scope it needs. A `write` scope includes `read` access to the same resource. Browser sessions have every scope.

    | Scope | Allows |
    |---|---|
    | `account:read` | Reading the account profile |
    | `account:write` | Changing account settings and the sleep schedule |
    | `calendars:read` | Reading calendars and events |
    | `calendars:write` | Creating, changing and deleting calendars and events |
    | `todos:read` | Reading todo lists, todos and checks |
    | `todos:write` | Creating, changing, completing and deleting todos |
...

...

    ## MCP

...

...
servers:
...
paths:
...
  /exports/ics:
    get:
      tags: [Exports]
      operationId: exportICS
      summary: Download calendars as iCalendar
      description: |
        Returns an iCalendar document with the chosen calendars, their recurring series and changed occurrences. Without `calendar_id` every calendar is included.

        Requires scope `exports:read`.
      x-scope: exports:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/ExportCalendars"
      responses:
        "200":
          description: An iCalendar document.
          headers: *okHeaders
          content:
            text/calendar:
              schema:
                $ref: "#/components/schemas/ICSDocument"
              example: |
                BEGIN:VCALENDAR
                VERSION:2.0
                END:VCALENDAR
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /exports/csv:
    get:
      tags: [Exports]
      operationId: exportCSV
      summary: Download calendars as CSV
      description: |
        Returns one row per event, with the columns Google and Outlook use. Recurring series appear once, as the rule rather than its occurrences.

        Requires scope `exports:read`.
      x-scope: exports:read
      security: *readAuth
      parameters:
        - $ref: "#/components/parameters/ExportCalendars"
      responses:
        "200":
          description: A CSV file.
          headers: *okHeaders
          content:
            text/csv:
              schema:
                $ref: "#/components/schemas/CSVDocument"
              example: "Subject,Start Date\nStandup,2026-09-07\n"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: cal_<prefix>_<secret>
      description: API token created in Settings → API tokens.
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-sid
      description: >-
        Browser session cookie, set by signing in. It is named `sid` when the server runs over plain HTTP.
        Unsafe methods also need the `X-CSRF-Token` header.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        CSRF token of the browser session, required together with `sessionCookie` on unsafe methods.
        Pages expose it in the `csrf-token` meta tag.
  parameters:
...
    ExportCalendars:
      name: calendar_id
      in: query
      required: false
      description: Only these calendars. Repeat the parameter or separate IDs with commas. Without it every calendar is included.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
          format: uuid
      example: ["0191f2c4-7a1b-7c3e-9a51-2b7d0e4f6a10"]
...
  headers:
...
    RateLimitLimit:
      description: Requests allowed per window.
      schema:
        type: integer
        minimum: 1
      example: 600
    RateLimitRemaining:
      description: Requests left in the current window.
      schema:
        type: integer
        minimum: 0
      example: 599
    RateLimitReset:
      description: Seconds until the full quota is available again.
      schema:
        type: integer
        minimum: 0
      example: 1
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        minimum: 1
      example: 12
    RequestID:
      description: Request identifier, also returned in error bodies as `request_id`.
      schema:
        type: string
      example: tQ6DNZRbQcyrEPPA
  responses:
...
    Unauthorized:
      description: The token or session is missing, invalid, revoked or expired.
      headers:
        WWW-Authenticate:
          description: Authentication scheme to use.
          schema:
            type: string
          example: 'Bearer realm="calendar"'
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Unauthorized
            status: 401
            detail: invalid api token
            instance: /api/v1/me
            request_id: tQ6DNZRbQcyrEPPA
    Forbidden:
      description: The token lacks the required scope, or a session request failed the CSRF or origin check.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            scope:
              summary: Missing scope
              value:
                type: "about:blank"
                title: Forbidden
                status: 403
                detail: "token lacks scope calendars:write"
                instance: /api/v1/calendars
                request_id: tQ6DNZRbQcyrEPPA
            csrf:
              summary: Missing CSRF token
              value:
                type: "about:blank"
                title: Forbidden
                status: 403
                detail: invalid csrf token
                instance: /api/v1/calendars
                request_id: tQ6DNZRbQcyrEPPA
...
    UnprocessableEntity:
      description: The request body is well formed but some fields are invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Unprocessable Entity
            status: 422
            detail: One or more fields are invalid.
            instance: /api/v1/calendars
            request_id: tQ6DNZRbQcyrEPPA
            errors:
              - field: color
                message: "must be a hex color like #3b6ea5"
    TooManyRequests:
      description: The rate limit was exceeded.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "about:blank"
            title: Too Many Requests
            status: 429
            detail: too many attempts, retry in 12s
            instance: /api/v1/me
            request_id: tQ6DNZRbQcyrEPPA
  schemas:
...
    ICSDocument:
      type: string
      description: An iCalendar document.
    CSVDocument:
      type: string
      description: A CSV file with one row per event.
...
    Problem:
      type: object
      description: Error details in the RFC 9457 format.
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          description: URI that identifies the error type. `about:blank` means the HTTP status explains it.
        title:
          type: string
          description: Short summary of the HTTP status.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Explanation specific to this request.
        instance:
          type: string
          description: Path of the request that failed.
        request_id:
          type: string
          description: Identifier of the request, the same as the `X-Request-ID` header.
        errors:
          type: array
          description: Invalid fields, present on `422` responses.
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      description: One invalid input field.
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
          description: Name of the field in the request.
        message:
          type: string
          description: What is wrong with the value.
...